one bug to fix related to updating labels per image on the screen correctly

+++ napari_bb_annotations/_key_bindings.py
@@ -6,12 +6,9 @@ import subprocess
 import numpy as np
 import pandas as pd
 from magicgui.widgets import ComboBox, Container, Table
-from napari import Viewer
 from PIL import Image, ImageDraw
+from .constants_lumi import BOX_ANNOTATIONS, LUMI_CSV_COLUMNS
 
-
-LUMI_CSV_COLUMNS = [
-    'image_id', 'xmin', 'xmax', 'ymin', 'ymax', 'label', 'prob']
 
 # create the GUI for selecting the values
 logger = logging.getLogger(__name__)
@@ -24,6 +21,24 @@ logging.basicConfig(
     datefmt="%Y-%m-%d:%H:%M:%S",
     level=logging.INFO,
 )
+
+
+def run_shell_cmd(cmd, fail_ok=False):
+    logger.info('running: {}'.format(cmd))
+    proc = subprocess.Popen(
+        cmd, shell=True, stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE)
+    (out, err) = proc.communicate()
+
+    out = out.decode('utf-8')
+    err = err.decode('utf-8')
+
+    if proc.returncode != 0 and not fail_ok:
+        logger.info('out: {}'.format(out))
+        logger.info('err: {}'.format(err))
+        raise AssertionError("exit code is non zero: %d" % proc.returncode)
+
+    return (proc.returncode, out, err)
 
 
 def check_bbox(bbox, width, height):
@@ -87,7 +102,7 @@ def get_bbox_obj_detection(bbox):
     return [x1, y1, x2, y2]
 
 
-def create_label_menu(shapes_layer, labels):
+def create_label_menu(shapes_layer, image_layer):
     """Create a label menu widget that can be added to the napari viewer dock
 
     Parameters:
@@ -104,7 +119,7 @@ def create_label_menu(shapes_layer, labels):
     """
     # Create the label selection menu
     label_property = "box_label"
-    label_menu = ComboBox(label='text label', choices=labels)
+    label_menu = ComboBox(label='text label', choices=BOX_ANNOTATIONS)
     label_widget = Container(widgets=[label_menu])
 
     def update_label_menu(event):
@@ -112,8 +127,12 @@ def create_label_menu(shapes_layer, labels):
         the current properties of the Shapes layer change
         """
         new_label = str(shapes_layer.current_properties[label_property][0])
-        if new_label != label_menu.value:
+        if new_label != label_menu.value and new_label in BOX_ANNOTATIONS:
             label_menu.value = new_label
+        elif (new_label in image_layer.metadata["new_labels"] and
+              new_label not in BOX_ANNOTATIONS and
+              new_label not in label_menu.choices):
+            label_menu.set_choice(new_label, new_label)
 
     shapes_layer.events.current_properties.connect(update_label_menu)
 
@@ -131,17 +150,16 @@ def create_label_menu(shapes_layer, labels):
     return label_widget
 
 
-def update_summary_table(shapes_layer, metadata):
-    labels = metadata["box_annotations"]
+def update_summary_table(shapes_layer, image_layer):
     box_labels = shapes_layer.properties["box_label"].tolist()
     split_dict = {
-        "data": [[box_labels.count(label)] for label in labels],
-        "index": tuple(labels),
+        "data": [[box_labels.count(label)] for label in BOX_ANNOTATIONS],
+        "index": tuple(BOX_ANNOTATIONS),
         "columns": ("c"),
     }
     table_widget = Table(value=split_dict)
     label_property = "box_label"
-    label_menu = ComboBox(label='text label', choices=labels)
+    label_menu = ComboBox(label='text label', choices=BOX_ANNOTATIONS)
 
     def update_table_on_label_change(event):
         """This is a callback function that updates the summary table when
@@ -151,22 +169,36 @@ def update_summary_table(shapes_layer, metadata):
         new_label = str(shapes_layer.current_properties[label_property][0])
         if new_label != label_menu.value:
             box_labels = shapes_layer.properties["box_label"].tolist()
-            data = [[box_labels.count(label)] for label in labels]
-            split_dict = {
-                "data": data,
-                "index": tuple(labels),
-                "columns": ("c"),
-            }
+            if new_label not in BOX_ANNOTATIONS:
+                new_labels = image_layer.metadata["new_labels"]
+                new_labels.append(new_label)
+                image_layer.metadata["new_labels"] = new_labels
+                index = sorted(BOX_ANNOTATIONS + [new_label])
+                data = [[box_labels.count(label)] for label in index]
+                split_dict = {
+                    "data": data,
+                    "index": tuple(index),
+                    "columns": ("c"),
+                }
+            else:
+                unique_labels = np.unique(
+                    shapes_layer.properties['box_label']).tolist()
+                data = [[box_labels.count(label)] for label in unique_labels]
+                split_dict = {
+                    "data": data,
+                    "index": tuple(unique_labels),
+                    "columns": ("c"),
+                }
             table_widget.value = split_dict
     shapes_layer.events.current_properties.connect(
         update_table_on_label_change)
 
     def update_table_on_coordinates_change(event):
         box_labels = shapes_layer.properties["box_label"].tolist()
-        data = [[box_labels.count(label)] for label in labels]
+        data = [[box_labels.count(label)] for label in BOX_ANNOTATIONS]
         split_dict = {
             "data": data,
-            "index": tuple(labels),
+            "index": tuple(BOX_ANNOTATIONS),
             "columns": ("c"),
         }
         table_widget.value = split_dict
@@ -176,9 +208,8 @@ def update_summary_table(shapes_layer, metadata):
     return table_widget
 
 
-@Viewer.bind_key('Shift-S')
 def save_bb_labels(viewer):
-    logger.info("Pressed key Shift-S")
+    logger.info("Pressed save bounding boxes, labels button")
     shape = viewer.layers["Shapes"]
     image = viewer.layers["Image"]
     metadata = viewer.layers["Image"].metadata
@@ -239,37 +270,45 @@ def save_bb_labels(viewer):
     df.to_csv(csv_path)
 
 
-@Viewer.bind_key('Shift-l')
 def load_bb_labels(viewer):
-    logger.info("Pressed key Shift-l")
+    logger.info("Pressed load bounding box, labels button")
+    if set(viewer.layers["Image"].metadata["loaded"]) == {True}:
+        return
     all_files = viewer.layers["Image"].metadata["all_files"]
     dirname = os.path.dirname(all_files[0])
-    df = pd.read_csv(os.path.join(
-        dirname, "bb_labels.csv"))
-    shape = viewer.layers["Shapes"]
-    bboxes = shape.data
-    labels = shape.properties["box_label"].tolist()
-    for index, row in df.iterrows():
-        x1 = row.xmin
-        x2 = row.xmax
-        y1 = row.ymin
-        y2 = row.ymax
-        label = row.label
-        image_id = row.image_id
-        z = all_files.index(image_id)
-        bbox_rect = np.array(
-            [[z, y1, x1], [z, y2, x1], [z, y2, x2], [z, y1, x2]]
-        )
-        bboxes.append(bbox_rect)
-        labels.append(label)
-    viewer.layers["Shapes"].data = bboxes
-    viewer.layers["Shapes"].properties["box_label"] = np.array(labels)
+    csv_path = os.path.join(dirname, "bb_labels.csv")
+    index_of_image = viewer.dims.current_step[0]
+    if (not viewer.layers["Image"].metadata["loaded"][index_of_image] and os.path.exists(csv_path)):
+        df = pd.read_csv(csv_path)
+        shape = viewer.layers["Shapes"]
+        bboxes = shape.data
+        labels = shape.properties["box_label"].tolist()
+        for index, row in df.iterrows():
+            x1 = row.xmin
+            x2 = row.xmax
+            y1 = row.ymin
+            y2 = row.ymax
+            label = row.label
+            image_id = row.image_id
+            z = all_files.index(image_id)
+            bbox_rect = np.array(
+                [[z, y1, x1], [z, y2, x1], [z, y2, x2], [z, y1, x2]]
+            )
+            bboxes.append(bbox_rect)
+            labels.append(label)
+            index_of_image = all_files.index(image_id)
+            viewer.layers["Image"].metadata["loaded"][index_of_image] = True
+            viewer.layers["Image"].metadata["inferenced"][index_of_image] = True
+        viewer.layers["Shapes"].data = bboxes
+        viewer.layers["Shapes"].properties["box_label"] = np.array(labels)
+    update_layers(viewer)
 
 
 def load_bb_labels_for_image(viewer):
-    logger.info("Loading already existing inference results for image")
+    logger.info("Loading existing inference results for image")
     all_files = viewer.layers["Image"].metadata["all_files"]
-    filename = all_files[viewer.dims.current_step[0]]
+    index_of_image = viewer.dims.current_step[0]
+    filename = all_files[index_of_image]
     dirname = os.path.dirname(all_files[0])
     df = pd.read_csv(os.path.join(dirname, "bb_labels.csv"))
     # Filter out the df for all the bounding boxes in one image
@@ -290,37 +329,26 @@ def load_bb_labels_for_image(viewer):
         )
         bboxes.append(bbox_rect)
         labels.append(label)
+    logger.info("labels are {}".format(labels))
     viewer.layers["Shapes"].data = bboxes
     viewer.layers["Shapes"].properties["box_label"] = np.array(labels)
+    viewer.layers["Image"].metadata["loaded"][index_of_image] = True
 
 
-@Viewer.bind_key('Shift-i')
 def run_inference_on_image(viewer):
-    logger.info("Pressed key Shift-i")
+    logger.info("Pressed button to run inference/prediction")
     image_layer = viewer.layers["Image"]
     metadata = image_layer.metadata
     all_files = metadata["all_files"]
-    filename = all_files[viewer.dims.current_step[0]]
+    index_of_image = viewer.dims.current_step[0]
+    filename = all_files[index_of_image]
     dirname = os.path.dirname(filename)
-    box_annotations = image_layer.metadata["box_annotations"]
     model = image_layer.metadata["model"]
 
-    basename = os.path.basename(filename)
     save_overlay_path = metadata["save_overlay_path"]
-    overlaid_save_name = os.path.join(
-        save_overlay_path,
-        "pred_{}".format(basename)
-    )
-    if os.path.exists(overlaid_save_name):
-        logger.info("Inference already ran on this image, calling load annotations instead")
-        load_bb_labels_for_image(viewer)
-    else:
+    csv_path = os.path.join(dirname, "bb_labels.csv")
 
-        labels_txt = os.path.join(dirname, "labels.txt")
-        with open(labels_txt, 'w') as f:
-            for index, label in enumerate(box_annotations):
-                f.write("{} {}\n".format(index, label))
-        csv_path = os.path.join(dirname, "bb_labels.csv")
+    if not viewer.layers["Image"].metadata["inferenced"][index_of_image]:
         # To not overwrite the existing csv, and lose the predictions per image
         # from last image
         if os.path.exists(csv_path):
@@ -329,22 +357,34 @@ def run_inference_on_image(viewer):
             df = pd.DataFrame(columns=LUMI_CSV_COLUMNS)
         csv_path_per_image = os.path.join(
             dirname, "bb_labels_{}.csv".format(os.path.basename(filename)))
-        subprocess.check_call(
-            'lumi predict {} --checkpoint {} -f {} --save_media_to {}'.format(
-                filename, model, csv_path_per_image, ), shell=True)
-        os.remove(csv_path_per_image)
-        frames = [df, pd.read_csv(csv_path_per_image)]
-        result_df = pd.concat(frames)
-        result_df.to_csv(csv_path)
-        logger.info("lumi prediction per image subprocess call completed ")
+        run_shell_cmd(
+            'lumi predict ' + '"{}"'.format(filename) +
+            ' --checkpoint ' + model +
+            ' -f ' + '"{}"'.format(csv_path_per_image) +
+            ' --save-media-to ' + save_overlay_path)
+        if os.path.exists(csv_path_per_image):
+            frames = [df, pd.read_csv(csv_path_per_image)]
+            os.remove(csv_path_per_image)
+            result_df = pd.concat(frames)
+            result_df.to_csv(csv_path)
+            logger.info("lumi prediction per image subprocess call completed ")
+            viewer.layers["Image"].metadata["inferenced"][index_of_image] = True
+        else:
+            logger.info("Prediction unsuccesful")
+        load_bb_labels_for_image(viewer)
+    else:
+        load_bb_labels_for_image(viewer)
 
 
-def update_layers(viewer, box_annotations):
-    logger.info("Pressed update layers button")
+def update_layers(viewer):
+    logger.info("Updating layers")
+    if viewer.layers["Image"].metadata["updated"]:
+        return
     shapes_layer = viewer.layers['Shapes']
+    image_layer = viewer.layers['Image']
     shapes_layer.mode = 'add_rectangle'
 
-    label_widget = create_label_menu(shapes_layer, box_annotations)
+    label_widget = create_label_menu(shapes_layer, image_layer)
     text_property = "box_label"
     text_color = 'green'
     text_size = 8
@@ -363,10 +403,43 @@ def update_layers(viewer, box_annotations):
     viewer.window.add_dock_widget(label_widget, area='right')
     table_widget = update_summary_table(
         viewer.layers["Shapes"],
-        viewer.layers["Image"].metadata)
+        viewer.layers["Image"])
     table_widget.min_width = 300
-    table_widget.min_height = 200
+    table_widget.min_height = 400
     table_widget.max_width = 300
-    table_widget.max_height = 200
-
+    table_widget.max_height = 400
+    viewer.layers["Image"].metadata["updated"] = True
     viewer.window.add_dock_widget(table_widget, area='right')
+
+
+def get_properties_table(current_properties):
+    split_dict = {
+        "data": [current_properties],
+        "index": tuple(["properties"]),
+        "columns": ("c"),
+    }
+    table_widget = Table(value=split_dict)
+    return table_widget
+
+
+def edit_bb_labels(viewer):
+    logger.info("Pressed edit labels for a bounding box button")
+    shapes_layer = viewer.layers["Shapes"]
+
+    current_properties = shapes_layer.current_properties['box_label'].tolist()
+    table_widget = get_properties_table(current_properties)
+
+    def on_item_changed(item):
+        # item will be an instance of `QTableWidgetItem`
+        # https://doc.qt.io/Qt-5/qtablewidgetitem.html
+        # whatever you want to do with the new value
+        new_label = item.data(table_widget._widget._DATA_ROLE)
+        new_label = new_label.capitalize()
+        current_properties = shapes_layer.current_properties
+        current_properties['box_label'] = np.asarray([new_label])
+        shapes_layer.current_properties = current_properties
+        table_widget.clear()
+
+    table_widget.native.itemChanged.connect(on_item_changed)
+
+    viewer.window.add_dock_widget(table_widget, area='left')
